requirements.txt
Flask
gunicorn
futures
farm-haystack